requirements.txt
streamlit

pandas

pillow